requirements.txt
aiohttp==3.10.11
aiosmtplib==3.0.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
blinker==1.9.0
boto3==1.37.27
botocore==1.37.27
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
coverage==7.7.1
cycler==0.12.1
databricks-sdk==0.47.0
debugpy==1.8.13
decorator==5.2.1
deprecated==1.2.18
executing==2.2.0
fastapi==0.115.11
flask==3.1.1
fonttools==4.56.0
gitdb==4.0.12
gitpython==3.1.44
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.6.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.18.1
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jupyter==1.0.0
jupyter_client==8.6.3
jupyter_core==5.8.0
kiwisolver==1.4.8
mako==1.3.9
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.22.0
mlflow-skinny==2.22.0
namex==0.0.8
nest-asyncio==1.6.0
notebook==7.3.1
numpy==2.1.3
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
optree==0.14.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.22.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2
pyzmq==26.3.0
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
scapy==2.4.5
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.39
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.0
termcolor==2.5.0
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.5.0
uvicorn==0.34.0
wcwidth==0.2.13
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zipp==3.21.0

aiosqlite
prometheus-client
